Update _index.md
Mostly just stylistic edits to try and improve readability - merge wherever you think it makes sense. A matter of taste in most cases. 

In one or two cases I felt an edit was required but was unsure of what to propose - e.g. do we "press" a pipeline? Feels like maybe this made better sense originally as "trykk" in Norwegian :) I chose select but maybe that's not right either