Test assert flag for attest pr commands (#588)
Our attest pr ... commands allow the use of an assert flag, which returns a non-zero exit code if there are no PRs for the given commit.
Although these were working fine, they were all missing tests to check this functionality.
This PR adds tests for this to attest pr azure, attest pr bitbucket and attest pr github.
This PR also ensures that any errors received from the server are output from these commands even if the assert flag is used, which was not the case previously.